#38 some refactoring; removed unused code
