+++ intellij-plugin/src/main/kotlin/pro/bilous/intellij/plugin/action/menu/VerifiedEvent.kt
@@ -2,14 +2,14 @@ package pro.bilous.intellij.plugin.action.menu
 
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.project.Project
-import pro.bilous.intellij.plugin.Migrations
+import pro.bilous.difhub.Migrations
 import java.lang.IllegalArgumentException
 
 class VerifiedEvent(val e: AnActionEvent) {
 	val project: Project = e.project ?: throw IllegalArgumentException("Project not found")
-	val basePath: String = project.basePath ?: throw IllegalArgumentException("Base path not found")
+	val projectPath: String = project.basePath ?: throw IllegalArgumentException("Base path not found")
 
 	init {
-		Migrations.movePropertyOrganizationFromCredentialsToSettings(basePath)
+		Migrations.movePropertyOrganizationFromCredentialsToSettings(projectPath)
 	}
 }
