+++ intellij-plugin/src/main/kotlin/pro/bilous/intellij/plugin/gen/CodeGenerator.kt
@@ -4,6 +4,7 @@ import pro.bilous.intellij.plugin.PathTools
 import com.intellij.openapi.vfs.VirtualFileManager
 import io.airlift.airline.*
 import org.openapitools.codegen.cmd.*
+import pro.bilous.difhub.Migrations
 import java.util.*
 
 class CodeGenerator {
